+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+
+ - Added support for optional arguments to `r.JS()`
+
 ## v2.1.2 - 2016-07-22
 
 ### Added